+++ packages/react-compass/src/dropdown/option/index.styles.ts
@@ -43,6 +43,21 @@ export const StyledIcon = styled('div', {
   overflow: 'hidden',
 })
 
+
+export const StyledFlag = styled('div', {
+  display: 'flex',
+  alignItems: 'center',
+  justifyContent: 'center',
+  flexShrink: 0,
+  width: '$6',
+  height: '$6',
+  overflow: 'hidden',
+  svg: {
+    width: '$6',
+    height: '$6',
+  }
+})
+
 export const StyledColor = styled('div', {
   display: 'flex',
   alignItems: 'center',