+++ packages/react-compass/src/otp-input/usePrevious.ts
@@ -13,3 +13,4 @@ function usePrevious<T>(value?: T) {
 }
 
 export default usePrevious
+